requirements.txt
pandas
numpy
psycopg2-binary
os
datetime
apache-airflow